fix(ui): SliderComponentのInstanceID型正常化・基底API活用および描画クラスのドキュメント刷新
- SliderComponent: handleObjectID_の型をintからuint64_tに変更し、32bit超での符号オーバーフロー・オブジェクト探索失敗を防止
- SliderComponent: プロパティ登録にRegisterGameObjectRefを適用し、Update内の多段階ポインタ参照をGetEngine()へ統一
- Documentation: Skybox.h, LineClass.h, Sprite.hの機械生成コメントを詳細な日本語Doxygenへ刷新

+++ project/IrufemiEngine/Framework/Component/UI/SliderComponent.h
@@ -49,7 +49,7 @@ public:
     /**
      * @brief ハンドル（ツマミ）となるGameObjectのIDを設定
      */
-    void SetHandleObjectID(int id);
+    void SetHandleObjectID(uint64_t id);
 
 private:
     bool CheckBounds(const Irufemi::Vector2& mousePos);
@@ -59,7 +59,7 @@ private:
     SpriteRendererComponent* backgroundSprite_ = nullptr;
     GameObject* handleObject_ = nullptr;
 
-    int handleObjectID_ = 0;                   // Prefabからの復元用
+    uint64_t handleObjectID_ = 0;              // Prefabからの復元用
     float value_ = 1.0f;                       // 0.0 ~ 1.0 の割合
     bool isDragging_ = false;                  // ドラッグ中か
     Irufemi::Vector2 hitboxScale_{1.0f, 1.0f}; // 当たり判定のスケール
